+++ R/discrete_table.R
@@ -1,3 +1,22 @@
+cat_long_all <- function(data, group = ".total", all_name = "All data") {
+  N <- nrow(data)
+  data_long <- pivot_longer(
+    data,
+    cols = all_of(group),
+    values_to = "by"
+  )
+  data_long <- group_by(data_long, by)
+  data_summ <- summarise(
+    data_long,
+    level = all_name,
+    N = .env[["N"]],
+    #value = paste(n(), paste0("(",digit1(100*n()/N),")")),
+    value = paste0("n = ", n()),
+    .groups = "drop"
+  )
+  pivot_wider(data_summ, names_from = "by")
+}
+
 #' Summarize categorical data
 #'
 #' @inheritParams pt_cont_wide
@@ -5,8 +24,6 @@
 #' of the full data in the output
 #' @param all_name label for full data summary
 #' @param nby number of unique levels for the `by` variable
-#' @param preshape if `TRUE`, returns summarized data prior to reshaping;
-#' this is intended for internal use
 #' @param wide `logical`; if `TRUE`, data frame will be returned in wide format;
 #' if `FALSE`, it will be returned in `long` format
 #'
@@ -17,7 +34,7 @@
 #' @export
 cat_data <- function(data, cols, by = ".total", panel = by,
                      summarize_all = TRUE, all_name = "All",
-                     wide = FALSE, nby = NULL, preshape = FALSE) {
+                     wide = FALSE, nby = NULL) {
 
   cols <- new_names(cols)
 
@@ -43,8 +60,6 @@ cat_data <- function(data, cols, by = ".total", panel = by,
 
   ans <- mutate(ans,name=names(cols)[.data[["name"]]])
 
-  if(preshape) return(ans)
-
   if(wide) {
     ans <- pivot_wider(
       ans,
@@ -53,6 +68,7 @@ cat_data <- function(data, cols, by = ".total", panel = by,
       names_sep = '.'
     )
   } else {
+    ans[["N"]] <- NULL
     ans <- pivot_wider(
       ans,
       names_from = by,
@@ -67,12 +83,24 @@ cat_data <- function(data, cols, by = ".total", panel = by,
 #'
 #' @inheritParams pt_cont_long
 #' @param span variable name for column spanner
+#' @param all_name_span table column name to use for data summaries across
+#' levels of `span` if it is provided
+#' @param summarize where to include a data summary across subgroups;
+#' use `none` to drop the summary from the table
 #' @param by use `span` argument instead
+#'
+#' @details
+#' The notes in this table are generated with [pt_cat_long_notes()].
 #'
 #' @export
 pt_cat_long <- function(data, cols, span  =  ".total",
-                        all_name = "All Groups", summarize_all = TRUE,
+                        all_name = " ",
+                        all_name_span = "Summary",
+                        summarize = c("both", "right", "top", "none"),
                         table = NULL, by = NULL) {
+
+  summarize <- match.arg(summarize)
+  summarize_all <- summarize != "none"
 
   has_span <- !missing(span)
 
@@ -80,13 +108,13 @@ pt_cat_long <- function(data, cols, span  =  ".total",
     warning("the 'by' argument was used; maybe you wanted 'span' instead?")
   }
 
-  if(span == ".total" & missing(all_name)) {
-    all_name <- "Summary"
+  if(span == ".total" & missing(all_name_span)) {
+    all_name_span <- "Summary"
   }
 
   cols <- new_names(cols, table = table)
 
-  data <- data_total_col(data, all_name)
+  data <- data_total_col(data, all_name_span)
 
   assert_that(length(span)==1)
   span <- new_names(span, table = table)
@@ -103,25 +131,22 @@ pt_cat_long <- function(data, cols, span  =  ".total",
     nby = nspan
   )
 
-  summarize_all <- summarize_all & nspan > 1
-
   if(summarize_all) {
-    all <- cat_data(
-      data,
-      cols = cols,
-      by = ".total",
-      nby = nspan,
-      all_name = all_name
-    )
-    names(all)[ncol(all)] <- bold_each(names(all)[ncol(all)])
-    ans <- left_join(ans, all, by = c("name", "level"))
-  }
-
-  if(exists("level", ans)) {
-    ans <- rename(ans, "\\ " = .data[["level"]])
+    if(has_span && summarize %in% c("right", "both")) {
+      all <- cat_data(
+        data,
+        cols = cols,
+        by = ".total",
+        nby = nspan,
+        all_name = all_name_span
+      )
+      all[["N"]] <- NULL
+      ans <- left_join(ans, all, by = c("name", "level"))
+    }
   }
 
   output_span <- NULL
+
   if(has_span) {
     output_span <- colgroup(names(span), unique(data[[span]]))
   }
@@ -130,14 +155,42 @@ pt_cat_long <- function(data, cols, span  =  ".total",
     data = ans,
     span = output_span,
     align = cols_center(.outer = 'l'),
-    col_rename = span,
+    cols_rename = span,
+    cols_blank = "level",
     panel = "name",
-    notes = "Summary is count (percent)"
+    notes = pt_cat_long_notes()
   )
 
-  out <- structure(out, class = "pmtable")
+  if(summarize %in% c("top", "both")) {
+    bot <- cat_long_all(data, unname(span))
+    if(summarize=="both") {
+      bot[[all_name_span]] <- paste0("n = ", nrow(data))
+    }
+    bot[["N"]] <- NULL
+    bot[["level"]] <- ""
+    out$cols_extra <- bot
+  }
+
+  out <- structure(out, class = c("pmtable", class(out)))
 
   return(out)
+}
+
+#' Return table notes for pt_cat_long
+#'
+#' See [pt_cat_long()].
+#'
+#' @param include_n if `TRUE`, add a note for `n`
+#' @param note_add additional notes to include
+#'
+#' @export
+pt_cat_long_notes <- function(include_n = TRUE, note_add = NULL) {
+  ans <- note_add
+  ans <- c(ans, "Summary is count (percent)")
+  if(isTRUE(include_n)) {
+    ans <- c(ans, "n: number of records summarized")
+  }
+  ans
 }
 
 #' Discrete data summary in long format
@@ -145,13 +198,19 @@ pt_cat_long <- function(data, cols, span  =  ".total",
 #' @inheritParams pt_cont_wide
 #' @param by a grouping variable for the summary; may be given as character
 #' vector or quosure
-#' @param summarize_all if `TRUE`, an overall summary will be appended to the
-#' table
+#' @param summarize where to put an all-data summary; choose `none` to omit the
+#' summary from the table
+#'
+#' @details
+#' The notes in this table are generated with [pt_cat_wide_notes()].
 #'
 #' @export
 pt_cat_wide <- function(data, cols, by = ".total", panel = by,
                         table = NULL, all_name = "All data",
-                        summarize_all = TRUE) {
+                        summarize = c("bottom", "none")) {
+
+  summarize <- match.arg(summarize)
+  summarize_all <- summarize != "none"
 
   has_by <- !missing(by)
   has_panel <- !missing(panel)
@@ -202,6 +261,10 @@ pt_cat_wide <- function(data, cols, by = ".total", panel = by,
 
   ans[[".total"]] <- NULL
 
+  if("N" %in% names(ans)) {
+    ans <- rename(ans, n = .data[["N"]])
+  }
+
   .panel <- rowpanel(NULL)
   if(has_panel) {
     .panel <- panel_data
@@ -218,10 +281,28 @@ pt_cat_wide <- function(data, cols, by = ".total", panel = by,
     align = cols_center(.outer = 'l'),
     cols_rename = c(.panel$col,by),
     panel = .panel,
-    notes = "Summary is count (percent)"
+    notes = pt_cat_wide_notes()
   )
 
-  out <- structure(out, class = "pmtable")
+  out <- structure(out, class = c("pmtable", class(out)))
 
   return(out)
+}
+
+
+#' Return table notes for pt_cat_wide
+#'
+#' See [pt_cat_wide()].
+#'
+#' @param include_n include the note for `n` column
+#' @param note_add additional notes to include
+#'
+#' @export
+pt_cat_wide_notes <- function(include_n = TRUE, note_add = NULL) {
+  ans <- note_add
+  ans <- c(ans, "Summary is count (percent)")
+  if(isTRUE(include_n)) {
+    ans <- c(ans, "n: number of records summarized")
+  }
+  ans
 }