+++ R/table-utils.R
@@ -66,6 +66,16 @@ italics_each <- function(x) {
 
 it_each <- italics_each
 
+split_bold <- function(x) {
+  if(!str_detect(x,fixed("..."))) {
+    return(bold_each(x))
+  }
+  x <- str_split(x, fixed("..."))
+  x <- map(x, bold_each)
+  x <- flatten_chr(x)
+  paste0(x, collapse = "...")
+}
+
 require_col <- function(data,col,context=NULL) {
   if(!exists(col,data)) {
     if(!is.null(context)) {
@@ -132,3 +142,13 @@ df_grepl_rows <- function(data, pattern, cols = names(data)) {
   seq(nrow(data)) %in% rows
 }
 
+paste_units <- function(cols, units) {
+  if(is.null(units) || length(cols)==0) return(cols)
+  unit_match <- match(cols,names(units))
+  col_match <- which(!is.na(unit_match))
+  unit_match <- na.omit(unit_match)
+  for(i in seq_along(col_match)) {
+    cols[[col_match[i]]] <- paste(cols[[col_match[i]]],units[[unit_match[i]]])
+  }
+  cols
+}