+++ tests/testthat/test-sizes.R
@@ -13,9 +13,6 @@ test_that("test-sizes-fontsize", {
 })
 
 test_that("test-sizes-rowspace", {
-  out <- inspect(ptdata(), sizes = tab_size(font = "tiny"))
-  expect_match(out$output[1], "\\tiny", fixed = TRUE)
+  out <- inspect(ptdata(), sizes = tab_size(row = 1.6))
+  expect_match(out$output, "renewcommand{\\arraystretch}{1.6}", fixed = TRUE, all=FALSE)
 })
-
-
-