+++ tests/testthat/test-table-utils.R
@@ -34,3 +34,12 @@ test_that("test-table-utils-stable_save", {
   expect_error(stable_save(x), "and the 'file' argument is missing")
 })
 
+test_that("table-utils paste units", {
+  cols <- LETTERS[c(2,5,4,3,1)]
+  units <- list(C = "pounds", X = "none", B = "mg", D = "kg", Z = "liters")
+  cols_new <- pmtables:::paste_units(cols, units)
+  expect_equal(
+    cols_new,
+    c("B mg", "E", "D kg", "C pounds", "A")
+  )
+})